+++ examples/src/main/java/com/demcha/examples/GenerateAllExamples.java
@@ -94,6 +94,7 @@ import com.demcha.examples.templates.cv.v2.CvModernV2Example;
 import com.demcha.examples.templates.cv.v2.CvMonogramSidebarExample;
 import com.demcha.examples.templates.cv.v2.CvNordicCleanExample;
 import com.demcha.examples.templates.cv.v2.CvPanelExample;
+import com.demcha.examples.templates.cv.v2.CvRuntimeModulesExample;
 import com.demcha.examples.templates.cv.v2.CvSidebarPortraitExample;
 import com.demcha.examples.templates.cv.v2.CvTimelineMinimalExample;
 import com.demcha.examples.templates.invoice.InvoiceCinematicFileExample;
@@ -139,6 +140,7 @@ public final class GenerateAllExamples {
         System.out.println("Generated: " + CvMonogramSidebarExample.generate());
         System.out.println("Generated: " + CvNordicCleanExample.generate());
         System.out.println("Generated: " + CvPanelExample.generate());
+        System.out.println("Generated: " + CvRuntimeModulesExample.generate());
         System.out.println("Generated: " + CvSidebarPortraitExample.generate());
         System.out.println("Generated: " + CvTimelineMinimalExample.generate());